+++ src/main/java/utils/Configuration.java
@@ -1,0 +1,6 @@
+package utils;
+
+public class Configuration {
+
+    public static final String BASE_URL = "https://qa-scooter.praktikum-services.ru";
+}
